+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 - Legend now works correctly again, for multiple layers with the same legend reference.
 - Fix map resized when a tool is as sidebar and defined with active:true in config.json
 - Issue #771: Fix issue of WMS-TimeLayer being displayed wrong and without TimeSlider, when set visivle through URLParameter
+- Layers are now also listed correctly in the `saveSelection` after the initial start of the portal.
 
 ---